tests/requirements.txt
# Test dependencies
selenium>=4.0.0
pyotp>=2.6.0
python-dotenv>=1.0.0
webdriver-manager>=4.0.0
